Fix use of certain empty sets from multiple optional arguments
The problem is that if a source aspect was missing for some set (like
because it came from an array_unpack, or was a function argument),
when ensure_source_rvar is run at the common binding point, it
unconditionally joined against the bare object table, which causes us
to lose our NULL values and get no rows returned instead, which means
our optional functions don't work.

We need to put an optional wrapping around it. Fortunately all the
optional wrapping machinery is easy to reuse here.

I also needed to new_empty_rvar to not ask to be joined against its
path_id, which will obviously never work.

Fixes #5979