Adjust explain tests to be less flaky.
There are tests that are intended to trigger indexes. It matters that
the specific indexes get used, but not necessarily the details of how
they get used, so whether the plan uses IndexScan or BitmapHeapScan is
secondary. Thus these tests should be flexible to allow either one of
those outcomes as valid.

Fixes: #5922