Do not stack overflow on long arrays and similar (#6004)
Allow JoinExpr to be flattened, instead of representing an N-ary join
as an N-deep binary tree.

Fixes #4380.

The SQL query performance still seems sometimes pretty bad, though,
for some of these tests. I think it is because pg really doesn't like
to get huge numbers of arguments.

We might consider capping how many constants we are willing to
extract.  And for this particular array case, extracting full arrays
would be good too.