Fix collisions between function parameters named id and the property (#6000)
We name function params their actual names, which can collide with the
`id` column in tables and produce bad results. It could maybe collide
with source and target in link tables too. It doesn't collide with
everything, because most columns get uuid names.

Qualify references to params with the function name.

Fixes #5935.