Fix multiplicity inference on operators
Currently we assume all operators with only one multi argument
are injective in the argument, which is not correct
(for example `0 * {2, 3}` gives back zero twice).

Fix this.

I've kept the old behavior for + and ++, the operators that we
previously tested preserved uniqueness, to avoid making too much of a
compatability break in the cases where the inference /was/ right.  I'm
not sure if this is worth doing, though it seems kind of reasonably
for ++...