Fix an infinite recursion when casting object->json in some cases (#6007)
When compiling binary operators, we try to constant fold them using
the static evaluator.

When we evaluate a type cast, we just check that the relevant types
support the casting, I think to make sure it is valid?

As part of that, we convert the schema types to python types, but the
function for that will infinite recurse if it hits a self link before
it hits something that makes it fail.

I fix this in two ways:
 * Prevent the infinite recursion in object_type_to_spec
 * Only do the cast checks for scalar types

I am probably going to send up a separate PR to fix this in a third
way by not doing the constant folding thing at all anymore.

Fixes #5569.