Inject the delta dump when we hit "not complete but no proposed"
It will probably always be inscrutable to users but I'll be able to
ask users for it when they are hitting certain nasty schema errors,
and it will be much easier for them to give it to me if they don't
need to restart the server with logging.

We can give the cli a flag to include the data.

I tested this using the bug in #5714.