Support marking config variables as "secret"
A secret pointer may not be accessed from a query, except via
functions and aliases defined in the same module as the pointer.

A lot of the work here is really in redacting the secret values along
a bunch of different paths.

We still need to decide what to do about dumps.

Fixes #5988.