Add an with_secrets header to dump to opt-in to dumping secret configs
The rust bindings don't support the new protocol version yet, and this
seemed like substantially the most straightforward way to do it. We
can try to do it differently if people really don't want to start
using headers again.

No tests for this, yet, since the CLI doesn't support it, and can't
support it until the server does, but I tested it manually.