Fix loading extensions with constraints via SDL (#6128)
There was a bug where ReferencedObjectCommand only worked if a fully
qualified name coudl be found for the object using modaliases to
resolve the module. If it was actually a nested object, though, then
we would discard that name.

Don't require that we be able to compute a wrong name before computing
the right one.

I don't have a test here, because the simplest test is loading the
auth extension via SDL, and that only works 50% of the time because we
don't have ext dependencies yet.