Support marking config variables as "secret" (#6106)
* Support marking config variables as "secret"

A secret pointer may not be accessed from a query, except via
functions and aliases defined in the same module as the pointer.

A lot of the work here is really in redacting the secret values along
a bunch of different paths.

Fixes #5988.

Secrets are omitted in dumps by default, but we add a new protocol
header to opt in to them.

The rust bindings don't support the new protocol version yet, and this
seemed like substantially the most straightforward way to do it. We
can try to do it differently if people really don't want to start
using headers again.

No tests for that, yet, since the CLI doesn't support it, and can't
support it until the server does, but I tested it manually.