Fix schema confusion with objects that shadow std objects
We were actually inferring the names of the *objects themselves*
as being std::Object when we declared a type named Object,
because we didn't understand that you never need to fall back
to looking in std when actually *declaring* the object.

Once that was fixed, there was also some impossibly confused code I
had apparently written that accidentally preferred names in std over
local ones as a result of a bad refactor.