Fix migration incomplete issues with functions that use multiple globals
We were treating used_globals as something that needed to perfectly
match (compcoef 0.0) when actually we had meant for it to be
ignored. And to make things worse, it was being compared in an order
sensitive way despite having a set somewhere in its dataflow.

Fixes an ancillary issue from #5641