Improve FTS weights
Limits what's accepted into `weight_category`.

Ideally, `with_options(weight_category)` would be an `anyenum`,
similar to language, so people can use their own enum,
with their own naming.
It would ignore the names of the enum variants and instead only
use their ordering which would have to match the ordering of
the weight values in the `fts::search`.

Right now, we cannot have it be an `anyenum`,
because the `anyenum` type would get conflated with `language: anyenum`.

Naive option would be to stick with `str` for now, but the future change
to enums would be backward in-compatible.

This PR adds `fts::Weight` enum that must be used in
`with_options(weight_category: fts::WeightCategory)`.
In the future, we can implement proper generics and change it
to `weight_category: anyenum`.

The category now also defaults to 'A',
instead of inheriting Postgres's behavior of defaulting to 'D'.

---

This PR also adjust `fts:search`'s `weights` parameter':
- it now defaults to `[1.0, 0.5, 0.25, 0.125]` (as agreed in the RFC),
- if less than four values are provided, it pads the array with zeros,
- if more than four values are provided, it ignores them,
- the four values of the array correspond to categories A, B, C and D, respectively.
  (instead of inheriting Postgres's behavior of D, C, B and A)