Disallow INSTANCE config of extension-defined config values
The main reason it was ever supported is that it was an implementation
stepping stone, since originally DATABASE didn't support CONFIGURE
INSERT. It doesn't really make sense, and it also doesn't work
currently, because if there is a config value from an extension set,
CONFIGURE INSTANCE operations will ISE if performed from other
databases, since they don't know how to serialize the config back to
json without a complete spec.

Drop the INSTANCE version of the extension config objects, too,
since it can never be populated.