Support DML on right hand side of coalesce expressions
The basic approach is to extract the pieces from the ?? and
rewrite them into:
    for optional x in (LHS,) union (
      {
	x.0,
	(for _ in (select () filter not exists x) union (RHS)),
      }
    )

Optional for is needed because the LHS needs to be bound in a for
in order to get put in a CTE and only executed once, but the RHS
needs to be dependent on the LHS being empty.

We hackily wrap the LHS in a 1-ary tuple and then project it back
out because the OPTIONAL FOR implementation doesn't properly
handle object-type iterators. OPTIONAL FOR relies on having a
non-NULL identity ref but objects use their actual id, which
will be NULL.

Follow-up to #6181. Like #6181, needs more testing.

One of the tests I wrote failed because of #6057, which I'm working on
fixing also.