Support listen_ports greater than 32767. (#6194)
The problem is that listen_port had a type of int16, but it really
should be an *unsigned* int16. We don't have those, so approximate.

I ran into this because it turns out that if the server is told to
automatically pick a port, *then* it works outside the int16
(... possibly because it is failing to reflect the value into config).
The `edb test --use-db-cache` flow would stop a server and then
restart it on its previous port, and it was failing.