Add a '_protected' flag to pointers
This is useful for config purposes.

Turns out that we *didn't* do a safety check for `__type__` on insert,
but would generate bogus SQL and ISE, so this fixes that.

Clean up some error messages in similar id related code, also.

I think we could just call it 'protected', though.