Fix DROP EXTENSION in a pretty heavy-handed way
The original approach for DROP EXTENSION used delta_schema and then
tried to apply it. I got rid of this in #5626 because delta_schema's
output is really only good for dumping as a DDL AST. But the existing
approach is pretty fragile, and the delta_schema()->AST->delta
pipeline is actually pretty well tested, so just do it.

This mostly restores the pre-#5626 approach, but adds a ->DDL->delta
stage.