Fix cardinality inference of polymorphic shape elements
We weren't properly accounting for the intersection when computing
cardinality of shape elements like `[is User].element`.
Note that fixing this can break queries, since now if you
write something like `Named { [is User].name }`, we will correctly
infer that the result card is optional and we will error, since
`Named.name` is `required`.

The core issue here is that we grab the original pointer from the
class named in the polymorphic query, and then make a subtype of it
that we attach to our real view. This results in us inheriting its
cardinality.

_normalize_view_ptr_expr is a pretty big tangle, and I have not
improved it today. I have an idea for a follow-up, though.

Fixes #6254.