Allow configuration of hosted auth flow type (#6247)
When returning the final auth token and identity, we need to choose between a PKCE flow, or an implicit flow. For now, we have this hard-coded to require a PKCE flow, but adding this configuration will allow developers to opt into the less secure implicit flow that we will eventually support.