Avoid using hashed postgres names for triggers
Triggers don't have the 64 character name limit, so don't worry about
it there.

I think this is the only place left in the *schema* where we hash
names to make them short enough, though we sometimes do it in the
queries we generate too. (Though that should be fixable as well.)