Fix behavior when a property appears in multiple splats (#6270)
* Avoid using hashed postgres names for triggers

Triggers don't have the 64 character name limit, so don't worry about
it there.

I think this is the only place left in the *schema* where we hash
names to make them short enough, though we sometimes do it in the
queries we generate too. (Though that should be fixable as well.)

* Fix behavior when a property appears in multiple splats

Currently one applies and then the other is ignored. This was revealed
by #6255, since now some of those queries will error.

We take the pointer from the ancestor class. If the types don't have
an ancestor relationship, we error.