Fix compilation of explicit ids in inserts that can be empty
Explicit ids in inserts that have nontrivial possibly empty
expressions would fail, because an interaction with optional wrapping
caused the id path to fall back to the identity of the source, which
the fallback rvar happily provided as NULL.

I ran across this bug in #3834, but hacked around it instead of fixing
it (the fix would have been totally different then, also), since
explicitly-specified `id`s were purely internal, so it wasn't
high-priority. Of course, two months later we added explicitly
specified ids.