Fix a dependence on order in a trigger test (#6280)
One of the parts of the assert_query_result shape used a list
incorrect instead of a set.

This has failed in testing *once* that we've noticed, while running
tests for the v4.0a3 release. I just reran the tests there without
fixing it in a3, but we'll cherry-pick this fix for b1.