Fix type name injection on computed backlinks without a shape
The support for computed backlinks in #5227 introduced a bug where
a computed backlink would have the material type as its target, not
the view into which `__tname__` had been injected, and so we wouldn't
report `__tname__` in our descriptor.

The root cause of this was that we were ignoring the ptr_target and
rederiving it from the source of the base pointer, which was losing
the shape. Not sure why we were doing that at all.

Fixes #5930