Fix renames of types with rewrites on them (#6293)
The mechanism for processing changes that affect expressions in the
schema (like renames to things used in expressions) was not correctly
initializing a context stack for the expression-containing objects it
modified. It had some custom code that created a context for the
outermost object it was processing, but this didn't work right for
more nested objects (an object type containing a pointer containing a
rewrite).

Fortunately init_delta_branch returns exactly the right context stack
to use, so just use it. (I think that this rename handling code
predates init_delta_branch doing this helpful thing.)