Medium-sized rework of normalization (#6299)
Rework the normalization pass to make it less likely to have bugs like
it didn't get noticed until triggers allowed putting DML in the schema
again.

It's already the case that *most* nodes have a totally generic
normalize implementation, but nodes that have any special handling
currently need fully special handling.

Instead, make the generic normalize function take a list of fields to
*skip*, which lets nodes have custom handling for the fields that need
it and then generically process the rest. This lets us merge all the
DML statements, also.

Then, registering the normalizer for ObjectRef allows us to drop the
handlers for Global and Path (which incidentally fixes normalization
of splats with type names in them).

The one snag is that Ptr and TupleElement use ObjectRef to refer to a
name that isn't allowed to be qualified. For now I have fixed this by
having a no-op normalizer for Ptr and then making TupleElement use Ptr
instead of ObjectRef, but I think I am going to follow up by making
Ptr just have a str instead of an ObjectRef, which is something that's
annoyed me for a while.