Store std schema and reflection schema in the same pickle (#6317)
The reflection schema is a thin modification of the std schema so them
sharing the same serialization container basically halves the memory
consumption of cached schema (saving about 20MiB).

This backport needs to do a bit more work than the original: if
std_and_reflection_schema is missing, we fall back to loading them
separately.