Return a WWW-Authenticate header when no authentication is attempted
This is how it is supposed to work, and if you are trying to use
urllib's actual basic auth features, the most obvious way (measured by
number of stack overflow answers suggesting it) of doing it *requires*
WWW-Authenticate.

Using urllib for one of the tests also has the benefit that we aren't
testing totally against ourselves.