Support auth for HTTP endpoints (#6352)
This makes all extension endpoints except for 'auth' will go through
authentication, since it seems like the safe default.

We add a new SIMPLE_HTTP transport method for HTTP endpoints.
(Unfortunately, edgedb+http already uses HTTP, even though
HTTP would be a better name for *this* one.)

We also add a new Password auth method that only works for
SIMPLE_HTTP, and which is the default method for SIMPLE_HTTP.
JWT authentication is also supported, but SCRAM is not.

This is a backwards incompatible change! Unless "Trust" is configured
as a method for HTTP transports (which it will be in dev instances),
HTTP endpoints will stop working when not authenticated to.

Fixes #6345.

Password authentication uses HTTP basic auth.

JWT authentication is done by adding a header of the form:
```
Authorization: bearer <JWT>
```

The username for JWT and Trust auth can be specified with the X-EdgeDB-User
header.