Add a hack to make migration deadlocks less likely in some common cases (#6379)
This is a partial mitigation of #6304.

There are a lot of opportunities for deadlocks between DDL
(which can take a lot of full locks on things) and
long-running queries.  One place this comes up is with views,
since recreating a view requires a lock. When adding things to
tables, our DDL must modify the table, then modify the view.
This poses a deadlock risk with queries, which nearly always
will access the view before accessing the table (since the table
is how they get to the view).

Put a hacky and hopefully temporary bandage around this particular
deadlock case by injecting a no-op ALTER VIEW before the first
ALTER of some table.  This ensures that DDL locks the inhview
before the real table, making the lock order match the typical
query lock order.