Add a test patch to test the new extension patching system
This is basically just lifted from #6398.

I tested by starting without this patch, clearing the
bootstrap cache, and running
`edb test tests/test_http_ext_auth.py --use-db-cache`

Then, applying this patch and running:
 * EDGEDB_DEBUG_BOOTSTRAP_CACHE_YOLO=1 edb test tests/test_http_ext_auth.py --use-db-cache
 * EDGEDB_DEBUG_BOOTSTRAP_CACHE_YOLO=1 edb test tests/test_http_ext_auth.py

which tests the case where the extension is already installed and the
case where it is not.