Support patching installed extensions and fix config patches
Installed extensions are tricky because they live in individual
database schemas and not in stdschema, so we need to check which
databases actually have the relevant extension and compile and run the
change against it directly.

Also, config patches were broken when we made ids deterministic for
stdlib and extension objects, because we stopped being able to rely on
uuid1 timestamps for order, so fell back to using the order in the
schema dicts, which isn't preserved by reflection. Fix this in a
pretty terrible way: query information_schema to extract the correct
order of columns in the view.

Also, fix deletion of `multi property` extension configs