Enable CI testing of patches to the auth extension
Have the patches testing script populate httpextauth with the
appropriate extensions, so that patching installed extensions will be
tested.

This required some adjusting of the test suite to make happen, since
BaseHttpExtensionTest would unconditionally create the extensions, but
we need to skip that if they already exist. It turns out that the base
DatabaseTestCase *also* has an EXTENSIONS property, and it populates
the extensions using a migration, which fits what we want perfectly.
Just ditch the whole `get_extension_name` bit.