Allow distribute EdgeDB server without bundled Postgres
This is not a part of the stable API, and may break in a later release.

Before this PR, an internal `_bundled_pg_version` depends on the result
of the command `pg_config`. This PR added an environment variable
`_EDGEDB_BUILDMETA_PG_VERSION` that also allows setting
`_bundled_pg_version` without calling `pg_config`.