Simplify ?? implementation when RHS is AT_MOST_ONE (#6529)
A lot of the code is dedicated to filtering out NULL results after the
COALESCE is done, but that is avoidable: if we make sure that
`nullable` is properly set on the expression, then this will be
handled automatically at higher levels of the compilation pipeline.

Once that simplification is made, the only difference from the ONE
case is that the RHS needs to be placed in a sub-SELECT, and we can
use set_as_subquery to do that. This lets us merge the two cases,
with the only difference being in how the RHS is compiled.