[3.x] Fix some interactions between FOR and optional values (#6526)
The primary bug this fixed in the original version had to do with
optional FOR and coalesced DML, but the bug was still possible to
trigger without that.

More generally, the issue here is that is that because of
pull_path_namespace, the outer iterator was being routed to an
optional wrapper that has the iterator in its non-NULL case, but which
would produce NULL in the NULL branch. We already have code to prevent
this sort of thing by putting visible paths in the path_id_mask in
update_scope (see #5575), but it was a little too limited.

I do still think that we probably should be doing this masking
unconditionally, but it causes lots of trouble with materialization
cases. I should follow up on that.

I also need to spend some time deciding whether the entire
pull_path_namespace system actually makes any sense at all
or whether it is just a hack we should hope to eliminate.