Generate better code for ?? with multi RHS or object type (#6532)
The multi RHS/object coalesce path basically replicated the optional
wrapper machinery. This shouldn't be necessary, since the LHS input is
*already* being placed in an optional wrapper, which means that we are
doing the optional machinery twice.

In practice, this means that the RHS gets executed unconditionally.
Change things to *not* repeat the optional wrapping, but to instead
compile the LHS on its own, and then do a union in which the RHS
filters on whether the LHS was NULL.
This lets the RHS only be executed conditionally.

Fixes #6528.