Use subqueries instead of optional wrappers when inserting
Currently when we have an optional single pointer that we are inserting,
we use optional wrappers to turn it into NULL when it is missing.

This is needlessly complex SQL output; use a subquery, like we do for
updates and shape selects. This sped an example I tried inserting 1m
objects with a potentially empty field from 13.5s -> 13s. Not sure
if that should be taken seriously, though.