Fix changing a pointer to be computed when there is a subtype
This was a regression introduced by #6542, which moved the deletion of
the pointer to happen *after* processing subcommands. It turns out
that that was correct, but in order for it to work right around
inheritance, we need to make sure that the propagation of alters into
child types occurs as "caused" actions, so that they happen after all
the processing for the parent objects themselves. (Including the
actual deletion of a pointer when it is changed to computed.)

Also, make sure that constraints get propagated when this change
happens.  That change then required a knock-on change to expression
propagation to avoid messing up error messages in that case.

Fixes #6542