Fix migrating between aliased and non-aliased computeds
This version is designed to be cherry-pickable to 4.x, so it doesn't
change how any of our schemas are represented, it just tries to
actually fully make the transition when we switch.

I'm going to do a follow-up where I rip all of this out and the cost
of schema changes.