Allow indexes that use user-defined functions to actually get hit (#6551)
In a 4.x regression, indexes/exclusive constraints that use
user-defined functions will have the function called with a json blob
that disables access policies as part of the globals argument.  This
causes the index expressions to not match, so the indexes won't get
used.

Users affected by the bug should either dump/restore or delete the
affected indexes and recreate them.