Drop legacy anonymous statement
The reason it was kept is explained in #3120, mitigated in #3814 #3837.
But the core issue that compiler assumes mutation of transaction state
for any compilation even by Parse message still exists. We can drop the
`_last_anon_compiled` because the client bindings are no longer sending
Parse + Execute for `commit` or `rollback` commands (which triggered two
server compilation of the same command) like edgedb/edgedb-python#337.
However, we still need to double check all bindings to make sure the
transaction control commands are using only one `Execute` message.