Get rid of the need to call qlparser.preload_spec()
It makes tests and tooling more annoying to do; the cost of supporting
lazy loading is an extra branch per parse, which is extremely low.
(I didn't want to fix edgedb-perf to call it, basically.)

I left the calls to preload_spec in the two production use cases, so
it happens at an explicitly predictable time.