Properly support @source/@target on multi-link constraints (#6585)
This is useful becase putting exclusive constraints on `(@source,
@lprop)` and `(@target, @lprop)` are both useful.  This is supposed to
be supported, but an overzealous checker blocks it for multi (which of
course is the only time when the @source one would be useful). Fix the
check.
Closes #6570 (by providing a proper way to accomplish what is desired).

While we are at it, disable @source and @target from being used in non
constraint/index contexts, since they are quite broken but
also not worth fixing. (I think @target is basically completely broken.
@source is also in bad shape but apparently at least sometimes works,
since two pgdump tests (of all things) use @source.)
Neither is documented at *all* currently.

Closes #1786.