Fix constant extraction's interaction with the new if-then-else
There is a problem that when there is a named argument and extracted
args, the extracted arguments get ordered based on the order they are
processed during compilation. This creates a bug when the ast nodes
aren't processed in source order, which they happen to not be for new
if-then-else.

Fixes #6494.