Avoid doing a join when injecting type ids
We have an optimization to allow skipping joins when all we want is
the id of a link, but it is suppressed if there is an access policy on
the target. Unsuppress it when the target is schema::ObjectType, since
there shouldn't be any user-visible links to an internal ObjectType
anyway, and it lets us avoid a join when injecting typeids.

Even better, this will in practice let us do type name injection
without a join also, since clients get a type mapping in the new
protocol.

Fixes #6599.