Fix some bugs involving union and coalescing of optional values (#6590)
There are a number of places that improperly handle a set that has
multiple NULL values in it:
 * INSERT on single values produced by ??
 * UPDATE on single values produced by ??
 * assert_single
 * limit/offset

The assert_single and limit case is a long standing bug that can be triggered
without much trouble. The INSERT/UPDATE bugs are new, and stem from
the new ?? on DML, since I think it that was the first thing in the
compiler backend where extra NULLS might be produced for something
with single cardinality.

Fix the long-standing cases by adding null checks, but fix the
coalesce cases by adding a null check at the producer side.
The rule then, is that there can be NULLs in any produced set
that can be empty, but a single set has to return at most one row
(so it can't have extra NULLs).


Fixes #6438.