[3.x] Fix some bugs involving union and coalescing of optional values (#6590)
The original 4.x/5.x version had a lot to do with DML coalescing.

There are a number of places that improperly handle a set that has
multiple NULL values in it:
 * assert_single
 * limit/offset

The assert_single and limit case is a long standing bug that can be triggered
without much trouble.

Fix the long-standing cases by adding null checks, but fix the
coalesce cases by adding a null check at the producer side.
The rule then, is that there can be NULLs in any produced set
that can be empty, but a single set has to return at most one row
(so it can't have extra NULLs).