Eliminate some uuid_generates in FOR cases where they aren't needed
Currently if we have a nested for loop like
```
for x in {1,2,3} union (for y in {3,4,5} union (x+y))
```
we will apply volatility refs when creating an iterator aspect
for y, which will cause the iterator aspect for x to be inserted
into the query.  This is all despite there not being anything at
all in the body that should require using the volatility refs.

Fix this by only creating the transient iterator aspect when
the FOR contains DML (or is optional).

This also means that for non-DML FORs, scalar iterators will fallback
to using row_number to force volatility (see get_volatility_ref)
while object type iterators will return to using their ids
(like before #6609).

(Wanting to fix this after thinking about it when dealing with the
auth performance situation is what sent me down on this whole rabbit
hole.)