Fix DML-containing FOR when the iterator is an object set with duplicates
DML-containing FOR loops get compiled into CTEs, and after the DML is
performed, the CTEs need to get joined back with the DML CTEs in order
to produce the output. Currently we do these joins on the 'identity'
of the iterator set; for non-object sets, we generate a transient
identity with uuid_generate_v4, and for object sets we ues the actual
object id.

That works ok if all the objects in the iterator set are unique, but
if they aren't we produce a lot of duplicate output rows.

To solve this, we need to use transient ids for object types too, but
overriding object identity with something transient causes problems:
if we need to actually join the object against some *actual* table, we
are in trouble, since we've overriden the identity.

Solve this by introducing a new 'iterator' aspect and distinguishing
between whether we want to join on an iterator aspect or a real identity
when making path bonds.

Fixes #6608