Fix another DROP DATABASE test flake
In #4629 we introduced a `drop_tb` testbase function for dropping
databases in tests that works around #4567, but #5032 added some new
tests that used `DROP DATABASE` directly.

This caused a test flake on a 3.x release build today; hopefully it
passes on the retry without me needing to cherry-pick this fix first.