Fix very slow global introspection query when there are lots of databases (#6633)
The introspection query was fetching annotations, which can never be
populated on databases and which was exhibiting some pretty
pathological quadratic time introspection queries. With 500 dbs,
it took me about 30s.

Also add a new knob to the patch system so that we can apply this in a
patch without it needing to do anything expensive on each database.